Remove the hard-coded "1" from multicore runs.
Allow the user to specify more than one run for multi-core benchmarks
while still defaulting to one run.  Multicore is a bit different in that
if you think about it, it's kind of "self averaging" since each run is
is the amalgamation of N concurrent processes.

Still, this lets users run more if they wish.